refactor: extract GHCi safety classifier
